Validate IRKs on all bond-store paths, fix advertising state on connect, widen capture counter
- Add is_valid_irk() check to all 5 bond-read paths that were publishing
  without validation (DISC_IMMEDIATE, ENC_CHANGE, DISC_DELAYED, ENC_LATE,
  RECONNECT_EXISTING)
- Set advertising_ = false in on_connect() to match documented state machine
  and publish switch state to HA
- Widen total_captures_ from uint8_t to uint32_t to prevent overflow in
  long-running continuous mode sessions
- Bump version to 1.5.9

+++ components/irk_capture/irk_capture.h
@@ -262,7 +262,7 @@ class IRKCaptureComponent : public Component {
     uint8_t capture_count;
   };
   std::vector<IRKCacheEntry> irk_cache_;  // Deduplication cache
-  uint8_t total_captures_ { 0 };          // Total IRKs captured this session
+  uint32_t total_captures_ { 0 };         // Total IRKs captured this session
   uint32_t last_publish_time_ { 0 };      // Last IRK publish timestamp
   uint32_t pairing_start_time_ { 0 };     // Global pairing timeout
